SignalLogger: Deprecate/replace threading hazard
Relevant to: #10228

Add storage modes to SignalLogger; allow the internal SignalLog member to
become a cache entry. Limit the data access methods on SignalLogger to the
legacy unsafe mode, and alternatively provide ergonomic lookup methods to get
the SignalLog from a context in the per-context mode.

Make SignalLog copyable and movable for compatibility with the cache.

Provide complete python bindings for SignalLogger, SignalLog, and the
(temporary) LogStorageMode enum.

Update numerous tests and examples.